pyproject.toml
[project]
name = "mle-bench-purple"
version = "0.1.0"
description = "Universal OpenAI-driven purple agent for MLE-Bench on AgentBeats"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.0.0",
    "openai>=1.75.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "lightgbm>=4.5.0",
    "xgboost>=2.0.0",
    "catboost>=1.2.7",
]

[project.optional-dependencies]
test = [
    "httpx>=0.28.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
